Rewrite setup for integration tests (simply supported center load)
