+++ tests/integration/simply-supported/point-loads/conftest.py
@@ -1,0 +1,6 @@
+import pytest
+
+
+@pytest.fixture(params=(-1000, -500, 500))
+def load(request):
+    return request.param
